Need to be fixed: adding manager

+++ trunk/SQLI_Gestion_RH_Web/src/main/java/ma/ensao/youmna/dao/TechnologieDao.java
@@ -21,6 +21,6 @@ public interface TechnologieDao {
 	
 	List<String> technologies();
 	
-	Map<String , Integer> getCountTechnologie();
+	Map<String , Long> getCountTechnologie();
 	
 }
